requirements.txt
# Core Framework
streamlit>=1.28.0

# Database
pymongo>=4.5.0

# GitHub Integration
pygithub>=2.1.1

# Data Visualization
plotly>=5.17.0
pandas>=2.1.0

# Environment Variables (optional, for .env file support)
python-dotenv>=1.0.0

# Additional Dependencies
requests>=2.31.0
